SOL-23: Harden attempt counter and tighten sequence test
Address review findings on #80:
- Widen the attempt counter to long so the clamp formula stays
  correct even under pathological reconnect counts.
- Collapse the delay-sequence test to a single action + single
  containsExactly assertion per the testing standard's one-when
  rule.

+++ prism/src/main/java/com/stablebridge/prism/infrastructure/grpc/ReconnectHandler.java
@@ -11,13 +11,13 @@ public class ReconnectHandler {
     private static final int MAX_ATTEMPT_EXPONENT = 4;
 
     private final ReentrantLock lock = new ReentrantLock();
-    private int attempt;
+    private long attempt;
 
     public long nextDelay() {
         lock.lock();
         try {
             attempt++;
-            var exponent = Math.min(attempt, MAX_ATTEMPT_EXPONENT);
+            var exponent = (int) Math.min(attempt, MAX_ATTEMPT_EXPONENT);
             var delay = RECONNECT_BASE_SECS * (1L << exponent);
             return Math.min(delay, RECONNECT_MAX_SECS);
         } finally {
